refactor: move standalones to functions.py; complete coverage
