+++ tests/unit/stream_processor/test_eventbridge_provider.py
@@ -6,7 +6,7 @@ from botocore import stub
 from pydantic import BaseModel
 
 from product.constants import XRAY_TRACE_ID_ENV
-from product.stream_processor.dal.events.base import build_events_from_models
+from product.stream_processor.dal.events.functions import build_events_from_models
 from product.stream_processor.dal.events.constants import EVENTBRIDGE_PROVIDER_MAX_EVENTS_ENTRY
 from product.stream_processor.dal.events.exceptions import ProductNotificationDeliveryError
 from product.stream_processor.dal.events.providers.eventbridge import EventBridge
